requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv
python-multipart
supabase
google-auth
google-auth-oauthlib
google-api-python-client
Pillow
PyPDF2
